Fix Nucleus Kits
remove positive speed attributes from the scout and medic kit in favour of walk-speed.

scout speed stays the same (speed was static as it came from armour).

medic speed increases from static value of 1.07 to 1.1 to make up for the loss of the sword's value of 1.1, which had made a total of 1.17 (medic now has a slower top speed, but isn't reliant on holding a certain item to achieve said value)