Change Circulus Turres to Circle Towers
This PR will change (Circulus Turres) and "Haunted Turres" (the seasonal reskin of "Circulus Turres") to "Circle Towers" and "Haunted Towers" respectively.
The reason for this name change is because the name was not liked by many, and it doesn't not make much sense anyways, so I thought it should be changed.