Gameplay changes for Trenches
Not pushing directly so feedback can be given. These are some small (but relevant) changes to Trenches. Recently had a match with 25+ players on, but it slowly (and steadily) dropped down to a 10v10 as it progressed.

20 minutes is painfully long for this map, considering the large distance you have to traverse to get the flag. The 15s recovery time is simply not enough if the flag is dropped near the enemy's spawn, flag net, or even their own "trench" (side). Have also slightly boosted the flag-holder's max health to make a "run for it" or escape more feasible. 

Additionally, regions & filters currently allow you to build horrible stalemate-y defenses like this: https://i.imgur.com/O6oJzs8.jpeg
I have not fixed that in this proposal but can totally do so if the rest agree or after the current time-related changes are tested.