Fourchette; make tools unbreakable & proto 1.4.0
Resolves #84

While testing Xounaut's pull request, I decided to convert the map to proto 1.4.0 for consistency. The platform to the wool rooms were slightly changed to add missing bedrock wall to the side  (one block wide tunnels), making it difficult to defend for teams. As a result, the map version has been bumped to 2.1.0. Everything has been tested and confirmed working.

Co-Authored-By: Xounaut <43223678+Xounaut@users.noreply.github.com>