Add "Pandora" by revinition
Pandora is a two-team 16v16 DTW map inspired by the movie, "Avatar". It has been tested by me, revinition, and mybeary and has been confirmed to be fully working. Rather than opening a pull request, I decided to commit it directly and test it on Warzone when the opportunity arises.